Update cipher.py
I used dictionary to solve the Caesar Cypher. The alphabet variable holds the letters (key) and their new value. For every letter the user inputs, the code will find the letter in the dictionary and add it to the variable secret message. Once its complete, the program will spit out the secret message.